Fixes and comments
