+++ src/csharp/Examples/WinForms/packages.config
@@ -6,4 +6,8 @@
   <package id="Microsoft.NetCore.Analyzers" version="2.9.3" targetFramework="net461" />
   <package id="Microsoft.NetFramework.Analyzers" version="2.9.3" targetFramework="net461" />
   <package id="StyleCop.Analyzers" version="1.1.118" targetFramework="net461" developmentDependency="true" />
+  <package id="System.Buffers" version="4.4.0" targetFramework="net461" />
+  <package id="System.Memory" version="4.5.3" targetFramework="net461" />
+  <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net461" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net461" />
 </packages>
