+++ src/csharp/K4a.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.572
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29111.141
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.Kinect.Sensor", "SDK\Microsoft.Azure.Kinect.Sensor.csproj", "{847B31D5-C253-4766-BF81-032F4670589D}"
 EndProject
@@ -27,6 +27,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{21E41070
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Files", "Solution Files", "{118BF8E1-C662-4852-955F-22DC68BB95CC}"
 	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
 		AzureKinectSensorSDK.ruleset = AzureKinectSensorSDK.ruleset
 		stylecop.json = stylecop.json
 		VersionInfo.cs = VersionInfo.cs
