+++ src/csharp/SDK/Device.cs
@@ -11,6 +11,8 @@ namespace Microsoft.Azure.Kinect.Sensor
     {
         private Device(NativeMethods.k4a_device_t handle)
         {
+            // Hook the native allocator and register this object.
+            // .Dispose() will be called on this object when the allocator is shut down.
             Allocator.Singleton.Hook(this);
 
             this.handle = handle;
